Moved validation logic to separate module
